requirements.txt
streamlit>=1.25.0
fpdf2>=2.7.0
